Update the TestCaseUtils class to change the return types of the getServerLdapPort, getServerLdapsPort, and getServerJmxPort methods to be int instead of long, and eliminate lots of now-unnecessary casts.
Also, add a new applyModifications(String... lines) convenience method that can
be used to easily apply a set of modifications to the server.

+++ opendj-sdk/opends/tests/unit-tests-testng/src/server/org/opends/server/api/PasswordValidatorTestCase.java
@@ -495,7 +495,7 @@ public class PasswordValidatorTestCase
     assertEquals(addOperation.getResultCode(), ResultCode.SUCCESS);
 
 
-    Socket s = new Socket("127.0.0.1", (int) TestCaseUtils.getServerLdapPort());
+    Socket s = new Socket("127.0.0.1", TestCaseUtils.getServerLdapPort());
     ASN1Reader r = new ASN1Reader(s);
     ASN1Writer w = new ASN1Writer(s);
     r.setIOTimeout(3000);
@@ -573,7 +573,7 @@ public class PasswordValidatorTestCase
     assertEquals(addOperation.getResultCode(), ResultCode.SUCCESS);
 
 
-    Socket s = new Socket("127.0.0.1", (int) TestCaseUtils.getServerLdapPort());
+    Socket s = new Socket("127.0.0.1", TestCaseUtils.getServerLdapPort());
     ASN1Reader r = new ASN1Reader(s);
     ASN1Writer w = new ASN1Writer(s);
     r.setIOTimeout(3000);
@@ -655,7 +655,7 @@ public class PasswordValidatorTestCase
     assertEquals(addOperation.getResultCode(), ResultCode.SUCCESS);
 
 
-    Socket s = new Socket("127.0.0.1", (int) TestCaseUtils.getServerLdapPort());
+    Socket s = new Socket("127.0.0.1", TestCaseUtils.getServerLdapPort());
     ASN1Reader r = new ASN1Reader(s);
     ASN1Writer w = new ASN1Writer(s);
     r.setIOTimeout(3000);
@@ -744,7 +744,7 @@ public class PasswordValidatorTestCase
     assertEquals(addOperation.getResultCode(), ResultCode.SUCCESS);
 
 
-    Socket s = new Socket("127.0.0.1", (int) TestCaseUtils.getServerLdapPort());
+    Socket s = new Socket("127.0.0.1", TestCaseUtils.getServerLdapPort());
     ASN1Reader r = new ASN1Reader(s);
     ASN1Writer w = new ASN1Writer(s);
     r.setIOTimeout(3000);
@@ -828,7 +828,7 @@ public class PasswordValidatorTestCase
     assertEquals(addOperation.getResultCode(), ResultCode.SUCCESS);
 
 
-    Socket s = new Socket("127.0.0.1", (int) TestCaseUtils.getServerLdapPort());
+    Socket s = new Socket("127.0.0.1", TestCaseUtils.getServerLdapPort());
     ASN1Reader r = new ASN1Reader(s);
     ASN1Writer w = new ASN1Writer(s);
     r.setIOTimeout(3000);
